Add error handling in release scripts

+++ set-version.bash
@@ -1,7 +1,12 @@
 #!/usr/bin/env bash
 
-echo "set version to $1"
-mv metadata.json temp.json
-jq -r --argjson V $1 '.version |= $V' temp.json > metadata.json
-rm temp.json
-echo "done"
+if [ $# -eq 0 ]
+  then
+    echo "No version argument given"
+  else
+    echo "set version to $1"
+    mv metadata.json temp.json
+    jq -r --argjson V $1 '.version |= $V' temp.json > metadata.json
+    rm temp.json
+    echo "done"
+fi
